Add partial caching support
